Remove SORF and budgeted PA.
