+++ epsilon/epsilon.h
@@ -33,44 +33,11 @@ float variance(const online_stats_t *s);
 // Perform in-place Fast Walsh-Hadamard transform.
 void FWHT(float *x, uint8_t nbits);
 
-// SORF is a pseudorandom matrix transformation based on fast Walsh-Hadamard
-// transform and sign flipping. Based on [4, 5]. Note that SORF is patented:
-// US2018/014145A1!
-void SORF(float *x, uint8_t nbits);
-
-// Repeat elements in x1 to form a vector x2 to increase input dimensions for
-// SORF.
-void SORF_repeat(float *x1, size_t n1, float *x2, size_t n2);
-
-// Kernels.
-typedef float (*kernel_t)(size_t i, size_t j);
-float squared_euclidean(kernel_t kernel, size_t a, size_t b);
-float squared_exponential_kernel(float length, float squared_dist);
-
-// Kernel projection used to implement kernel passive-aggressive algorithms.
-typedef struct {
-	float *alpha;
-	size_t num_alpha;
-	kernel_t kernel;
-} KP_t;
-
-// Apply kernel projection to input x1.
-float KP_apply(KP_t *km, size_t x1);
-
-// Get n-th idle (i.e. alpha=0) support vector of kernel projection.
-size_t KP_find_idle(const KP_t *km, size_t n);
-
-// Passive-aggressive parameters, see [6].
+// Passive-aggressive parameters, see [4].
 typedef struct {
 	float C;   // Perform aggressive updates with high C.
 	float eps; // Size of insensitive band for regression.
 } PA_t;
-
-// Perform kernel PA regression [1]. Target y can be NAN for inference.
-float KPA_regress(KP_t *km, const PA_t pa, size_t xi, float y);
-
-// Perform budgeted kernel PA regression [2]. Target y can be NAN for inference.
-float BKPA_regress(KP_t *km, const PA_t pa, size_t xi, float y);
 
 /*
 # References
@@ -84,18 +51,6 @@ float BKPA_regress(KP_t *km, const PA_t pa, size_t xi, float y);
 [3] Welford, B. P. "Note on a method for calculating corrected sums of
     squares and products." Technometrics 4.3 (1962): 419-420.
 
-[4] Felix, X. Yu, et al. "Orthogonal random features." Advances in Neural
-    Information Processing Systems. 2016.
-
-[5] Choromanski, Krzysztof, and Vikas Sindhwani. "Recycling randomness
-    with structure for sublinear time kernel expansions." International
-    Conference on Machine Learning. 2016.
-
-[6] Crammer, K. et al. “Online Passive-Aggressive Algorithms.” J. Mach. Learn.
+[4] Crammer, K. et al. “Online Passive-Aggressive Algorithms.” J. Mach. Learn.
     Res. (2003).
-
-[7] Wang, Zhuang, and Slobodan Vucetic.  "Online passive-aggressive
-    algorithms on a budget." Proceedings of the Thirteenth International
-    Conference on Artificial Intelligence and Statistics.  JMLR Workshop and
-    Conference Proceedings, 2010.
 */
